race conditions resolved
